Ref [TASK] [4595,4596,4597,4598,4599,4600,4601,4602,4603,4631,4632,4633,4634] chức năng liên quan đến ShoppingCart và Order

+++ src/main/java/com/ra/project/model/dto/request/ProductRequest.java
@@ -29,7 +29,7 @@ public class ProductRequest {
     private Double unitPrice;
     @Column(name = "stock_quantity")
     @Min(1)
-    private int stockQuantity;
+    private Long stockQuantity;
     @Column(name = "image")
     private String image;
     private Long categoryId;
